Added:Post-middleware-with-userSchema-and-productSchema

+++ backend/src/auth_module/controllers.ts
@@ -5,16 +5,13 @@ import jwt from 'jsonwebtoken'
 import dotenv from 'dotenv'
 dotenv.config()
 import appRes from "../utils/appRes"
-import userModel from '../models/user'
+import User from '../models/user'
 import adminEmails from "../utils/adminEmails"
 import isValidEmail from '../utils/isValidEmail';
 import path from 'path';
 import { deleteFile, processImage } from "../utils/imageProcessor"
 import TryCatch from "../middlewares/tryCatch"
 import { rm } from "fs"
-
-
-
 
 export const signup = async(req:Request, res:Response,next:NextFunction)=>{  
         const payload = req.body
@@ -35,7 +32,7 @@ export const signup = async(req:Request, res:Response,next:NextFunction)=>{
         }
 
         try {  
-            const emailExists = await userModel.findOne({email: payload.email})
+            const emailExists = await User.findOne({email: payload.email})
             if (emailExists){
                 if(avatar)rm(avatar.path,()=>{console.log('avatar path deleted')}) 
                 return next(appErr(`${payload.email} email is exists. Try another`,401))
@@ -51,7 +48,7 @@ export const signup = async(req:Request, res:Response,next:NextFunction)=>{
             payload.password = hashedPass;
             payload.answer = hashedAnswer;
             payload.role = getRole(email);
-            const user = new userModel(payload)
+            const user = new User(payload)
             await user.save();
             if(req.file){ 
                 const filename = await processImage({ 
@@ -60,32 +57,26 @@ export const signup = async(req:Request, res:Response,next:NextFunction)=>{
                     imgWidth: 100,
                     imgQuality: 80
                 })
-
                 user.avatar = path.join('./public/avatars', filename);
                 await user.save();
-
                 // Clean up temporary file after processing
                 deleteFile(path.join('./temp', req.file.filename));
             }
             user.password = undefined
             user.answer = undefined
             appRes(res,201,'','Registration success',{user})
-
         } catch (e:any) {
-            if (req.file) {
-                deleteFile(path.join('./temp', req.file.filename)); // Clean up on error
-            }
+            if (req.file){deleteFile(path.join('./temp', req.file.filename))} // Clean up on error
             return next(appErr(e.message,500))
         }
-    
     }
 
 export const signin = TryCatch( 
     async(req:Request,res:Response, next:NextFunction)=>{ 
         const {email,password} = req.body
-        if(!email || !password) return next(appErr('email and password are required',400))
+        if(!email || !password)return next(appErr('email and password are required',400))
         if(!isValidEmail(email))return next(appErr('Invalid email format',400)) 
-        const user = await userModel.findOne({email})
+        const user = await User.findOne({email})
         const isMatch = await bcrypt.compare(password, user!.password!)
         if(!isMatch)return next(appErr('Invalid Credentials',401))
         const secretKey = process.env.JWT_SECRET;
